refactor(app): split game_map_area_state_logic god file by concern
Splits the 1024-NCL god file from #2575 finding 12 / proposed work item
11 into three dedicated modules while keeping every public call site
working through thin static forwarders on `GameMapAreaStateLogic`:

- `game_map_area_civilian_draft_projection.dart` —
  `GameMapAreaCivilianDraftProjection.project` plus its private
  `_ProjectedCivilianUnit` helper and civilian-icon priority helpers.
- `game_map_area_fleet_draft_projection.dart` —
  `GameMapAreaFleetDraftProjection.project` plus the private
  `_FleetTileProj` helper.
- `game_map_area_province_action_states.dart` —
  `GameMapAreaProvinceActionStates.prospect` / `.explore` /
  `.buildImprovement` / `.buildExploreEligibleTileKeyCache` plus the
  hidden-state constants. Drive-by: replaces the remaining linear
  `for (final p in game.players)` lookup in build-improvement with
  `game.playerById`, aligning with the `app_lib_player_by_id_lookup`
  policy carried by PR #2581.

`game_map_area_state_logic.dart` shrinks from 1024 → 357 lines and now
just owns the work-target cache filtering / order-assignment helpers
plus thin forwarders that keep existing callers and tests, plus the
SPEC anchor in `SPEC/program/order-suggestions.md` § Authoritative
pipeline, unchanged.

Tests: existing 84 cases in
`game_map_area_state_logic_test_part1..3_test.dart` and the cache /
campaign suites still exercise the public API end-to-end and remain
green. Added `game_map_area_modular_split_test.dart` (9 cases) that
calls the new module entry points directly and asserts forwarder ↔
direct-call parity for civilian / fleet projection and the three
province action states, including negative malformed-tile-key paths.

Refs #2575
